+++ Assets/ViveGrip/Scripts/ViveGrip_Interactable.cs
@@ -2,8 +2,15 @@ using UnityEngine;
 
 [DisallowMultipleComponent]
 public class ViveGrip_Interactable : MonoBehaviour {
+  [Tooltip("Should the highlighting on this object be turned off?")]
+  public bool disableHighlight = false;
+
   void Start() {
     ViveGrip_Highlighter.AddTo(gameObject);
+  }
+
+  void Update() {
+    GetComponent<ViveGrip_Highlighter>().enabled = !disableHighlight;
   }
 
   // These are called this on the scripts of the attached object and children of the controller: